+++ lang/en/theme_training.php
@@ -1,21 +1,23 @@
 <?php
-// This file is part of Moodle - http://moodle.org/
+// This file is part of Training Theme for Moodle - http://moodle.org/
 //
-// Moodle is free software: you can redistribute it and/or modify
+// Training is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
 // the Free Software Foundation, either version 3 of the License, or
 // (at your option) any later version.
 //
-// Moodle is distributed in the hope that it will be useful,
+// Training Theme for Moodle is distributed in the hope that it will be useful,
 // but WITHOUT ANY WARRANTY; without even the implied warranty of
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 // GNU General Public License for more details.
 //
 // You should have received a copy of the GNU General Public License
-// along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
+// along with Training Theme for Moodle.  If not, see <http://www.gnu.org/licenses/>.
+
+// Strings for component 'theme_training', language 'en'.
 
 /**
- * Strings for component 'theme_training', language 'en'
+ * Training theme is based on the Moodle Clean theme.
  *
  * @package   theme_training
  * @copyright 2013 Moodle, moodle.org, 2017 TNG Consulting Inc. - www.tngconsulting.ca
@@ -30,41 +32,49 @@ $string['choosereadme'] = '
 </div>
 <div class="well">
 <h3>About</h3>
-<p>Training is a custom version of the Moodle Clean theme, which is based on Bootstrap 2.3. It inherits styles and renderers from its parent theme.</p>
-<h3>Parents</h3>
-<p>This theme is based upon the Bootstrap and Clean themes, which were created for Moodle 2.5, with the help of:<br>
+<p>Training is a simple and easy modern looking theme uncluttered by lots of options which is great for beginners or when you just need to get something that looks great out the door quickly.</p>
+<h3>Human made</h3>
+<p>Training is based upon the Bootstrap and Clean themes, HTML5/Bootstrap 2.3 based responsive themes, which was initially created for Moodle 2.5 with the help of:<br>
 Stuart Lamour, Mark Aberdour, Paul Hibbitts, Mary Evans, Bas Brands and David Scotson.</p>
 <h3>Theme Credits</h3>
-<p>Authors: Michael Milette<br>
+<p>Authors: TNG Consulting Inc. - Michael Milette<br>
 Website: <a href="http://www.tngconsulting.ca">www.tngconsulting.ca</a>
 </p>
 <h3>Report a bug:</h3>
 <p><a href="https://www.github.com/michael-milette/moodle-theme_training/issues">https://www.github.com/michael-milette/moodle-theme_training/issues</a></p>
 </div></div>';
 
+$string['pluginname'] = 'Training';
 $string['configtitle'] = 'Training';
 
-$string['customcss'] = 'Custom CSS';
-$string['customcssdesc'] = 'CSS rules added to this textarea will be reflected in every page, making for easier customization of this theme.';
+$string['favicon'] = 'Favicon';
+$string['favicondesc'] = 'The favicon appears in the tabs of some web browsers and may also appear on the desktop if you create a shortcut icon. Favicons should be 16px X 16px in PNG or ICO file format.<br>Default: Moodle icon.';
 
-$string['footnote'] = 'Footnote';
-$string['footnotedesc'] = 'Content add to this textarea will be displayed in the footer throughout your Moodle site.';
+$string['sitelogo'] = 'Site logo';
+
+$string['banner'] = 'Banner';
+$string['bannerdesc'] = 'Check this box if the logos that follow are actually full width banner images. These should be a minimum of 1920px wide.<br>Tip: If you want to have your logo and the banner, incorporate your logo into the banner image. Want text on the banner? Include it on the banner and be sure to apply the same text for the full site name as it is used for the ALT text of the image.';
+
+$string['logo'] = 'Logo';
+$string['logodesc'] = 'The standard logo is only displayed in the header of the front page and login page.<br>Tip: Logos can be up to 700px wide. There is no limit for the height through it might start to look weird if it is too tall.<br>If a logo is not added here, any logo in Appearance > Logos will be used. If that does not exist, the Site name will appear instead.';
+
+$string['smalllogo'] = 'Small logo';
+$string['smalllogodesc'] = 'The small logo is displayed in the header of all pages other than the front page and login page.<br>Tip: If you want the logo to be the same size on all pages, upload the same size logo for both the <strong>Logo</strong> and <strong>Small logo</strong>.<br>Default: If a logo is not added here, the compact logo in Appearance > Logos will be used. If that does not exist, the Site name will appear instead.';
 
 $string['invert'] = 'Invert navbar';
 $string['invertdesc'] = 'Swaps text and background for the navbar at the top of the page between black and white.';
 
-$string['logo'] = 'Logo';
-$string['logodesc'] = 'The standard logo is only displayed in the header of the front page and login page.<br />If a logo is not added here, any logo in Appearance > Logos will be used.';
-
-$string['smalllogo'] = 'Small logo';
-$string['smalllogodesc'] = 'The small logo is displayed in the header of all pages other than the front page and login page. If there is a header logo for the front page and login page, the small logo is not displayed on these pages. If a logo is not added here, the compact logo in Appearance > Logos will be used.';
-
-$string['favicon'] = 'Favicon';
-$string['favicondesc'] = 'The favicon appears in the tabs of some web browsers and may also appear on the desktop if you create a shortcut icon.<br />If a favicon is not added here, the default Moodle icon will be will be used.';
-
-$string['pluginname'] = 'Training';
-
 $string['region-side-post'] = 'Right';
 $string['region-side-pre'] = 'Left';
 
-$string['sitelogo'] = 'Site logo';
+$string['gridview'] = 'Masonry (grid style)';
+$string['gridviewdesc'] = 'Check this box if the course lists should be diplayed in a masonry style. When not checked, the default is full width where the course featured image is to the left of the course summary. Note that this option does not affect the <strong>Combo list</strong> listings.';
+
+$string['hidefrontsitepages'] = 'Hide site pages on front page';
+$string['hidefrontsitepagesdesc'] = 'Check this box if you want to hide the site pages on the front page. This enables you to create publicly accessible pages without enabling guest user access.';
+
+$string['footnote'] = 'Footnote';
+$string['footnotedesc'] = 'Content add to this textarea will be displayed in the footer throughout your Moodle site.<br>Tip: You can create columns in your footer by using <a href="https://getbootstrap.com/2.3.2/scaffolding.html"</a>Bootstrap Scaffolding</a>.';
+
+$string['customcss'] = 'Custom CSS';
+$string['customcssdesc'] = 'CSS rules will be reflected in every page, making for easier customization of this theme.';