+++ layout/embedded.php
@@ -18,9 +18,11 @@
  * The embedded layout.
  *
  * @package   theme_training
- * @copyright 2013 Moodle, moodle.org
+ * @copyright 2013 Moodle, moodle.org, 2017 TNG Consulting Inc. - www.tngconsulting.ca
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
+
+defined('MOODLE_INTERNAL') || die;
 
 echo $OUTPUT->doctype() ?>
 <html <?php echo $OUTPUT->htmlattributes(); ?>>