+++ settings.php
@@ -1,27 +1,21 @@
 <?php
-// This file is part of Moodle - http://moodle.org/
+// This file is part of Training Theme for Moodle - http://moodle.org/
 //
-// Moodle is free software: you can redistribute it and/or modify
+// Training is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
 // the Free Software Foundation, either version 3 of the License, or
 // (at your option) any later version.
 //
-// Moodle is distributed in the hope that it will be useful,
+// Training Theme for Moodle is distributed in the hope that it will be useful,
 // but WITHOUT ANY WARRANTY; without even the implied warranty of
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 // GNU General Public License for more details.
 //
 // You should have received a copy of the GNU General Public License
-// along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
+// along with Training Theme for Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
 /**
- * Training theme, a Clean child theme.
- *
- * DO NOT MODIFY THIS THEME!
- * COPY IT FIRST, THEN RENAME THE COPY AND MODIFY IT INSTEAD.
- *
- * For full information about creating Moodle themes, see:
- * http://docs.moodle.org/dev/Themes_2.0
+ * Training theme is based on the Moodle Clean theme.
  *
  * @package   theme_training
  * @copyright 2013 Moodle, moodle.org, 2017 TNG Consulting Inc. - www.tngconsulting.ca
@@ -32,17 +26,25 @@ defined('MOODLE_INTERNAL') || die;
 
 if ($ADMIN->fulltree) {
 
-    // Invert Navbar to dark background.
-    $name = 'theme_training/invert';
-    $title = get_string('invert', 'theme_training');
-    $description = get_string('invertdesc', 'theme_training');
+    // Favicon file setting.
+    $name = 'theme_training/favicon';
+    $title = get_string('favicon', 'theme_training');
+    $description = get_string('favicondesc', 'theme_training');
+    $setting = new admin_setting_configstoredfile($name, $title, $description, 'favicon');
+    $setting->set_updatedcallback('theme_reset_all_caches');
+    $settings->add($setting);
+
+    // Use banners instead of logos.
+    $name = 'theme_training/banner';
+    $title = get_string('banner', 'theme_training');
+    $description = get_string('bannerdesc', 'theme_training');
     $setting = new admin_setting_configcheckbox($name, $title, $description, 0);
     $setting->set_updatedcallback('theme_reset_all_caches');
     $settings->add($setting);
 
     // Logo file setting.
     $name = 'theme_training/logo';
-    $title = get_string('logo','theme_training');
+    $title = get_string('logo', 'theme_training');
     $description = get_string('logodesc', 'theme_training');
     $setting = new admin_setting_configstoredfile($name, $title, $description, 'logo');
     $setting->set_updatedcallback('theme_reset_all_caches');
@@ -56,11 +58,36 @@ if ($ADMIN->fulltree) {
     $setting->set_updatedcallback('theme_reset_all_caches');
     $settings->add($setting);
 
-    // Favicon file setting.
-    $name = 'theme_training/favicon';
-    $title = get_string('favicon', 'theme_training');
-    $description = get_string('favicondesc', 'theme_training');
-    $setting = new admin_setting_configstoredfile($name, $title, $description, 'favicon');
+    // Invert Navbar to dark background.
+    $name = 'theme_training/invert';
+    $title = get_string('invert', 'theme_training');
+    $description = get_string('invertdesc', 'theme_training');
+    $setting = new admin_setting_configcheckbox($name, $title, $description, 0);
+    $setting->set_updatedcallback('theme_reset_all_caches');
+    $settings->add($setting);
+
+    // Display courses using grid view.
+    $name = 'theme_training/hidefrontsitepages';
+    $title = get_string('hidefrontsitepages', 'theme_training');
+    $description = get_string('hidefrontsitepagesdesc', 'theme_training');
+    $setting = new admin_setting_configcheckbox($name, $title, $description, 0);
+    $setting->set_updatedcallback('theme_reset_all_caches');
+    $settings->add($setting);
+
+    // Display courses using grid view.
+    $name = 'theme_training/gridview';
+    $title = get_string('gridview', 'theme_training');
+    $description = get_string('gridviewdesc', 'theme_training');
+    $setting = new admin_setting_configcheckbox($name, $title, $description, 0);
+    $setting->set_updatedcallback('theme_reset_all_caches');
+    $settings->add($setting);
+
+    // Footnote setting.
+    $name = 'theme_training/footnote';
+    $title = get_string('footnote', 'theme_training');
+    $description = get_string('footnotedesc', 'theme_training');
+    $default = '';
+    $setting = new admin_setting_confightmleditor($name, $title, $description, $default);
     $setting->set_updatedcallback('theme_reset_all_caches');
     $settings->add($setting);
 
@@ -73,12 +100,4 @@ if ($ADMIN->fulltree) {
     $setting->set_updatedcallback('theme_reset_all_caches');
     $settings->add($setting);
 
-    // Footnote setting.
-    $name = 'theme_training/footnote';
-    $title = get_string('footnote', 'theme_training');
-    $description = get_string('footnotedesc', 'theme_training');
-    $default = '';
-    $setting = new admin_setting_confightmleditor($name, $title, $description, $default);
-    $setting->set_updatedcallback('theme_reset_all_caches');
-    $settings->add($setting);
 }