+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [0.2.0] - 2017-11-06
+### Added
+- Option to use a banner instead of a logo;
+- Option to display of course list in grid view;
+- Option to hide site pages on front page.
+
 ## [0.1.0] - 2017-11-05
 ### Added
 - Initial public ALPHA release on GitHub.